backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
requests==2.32.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.2
psycopg2-binary==2.9.9
